+++ src/components/ui/textEditor/editor.js
@@ -1,7 +1,7 @@
 // @ts-nocheck
 import { ThemeHelper } from '../../../theme/theme';
-import { CustomEventService, IdService, LoggerService, HTMLService } from '../../../services';
-import { NumberService, JSONService, ObjectService } from '../../../services/utils';
+import { CustomEventService, IdService, LoggerService, HTMLService, StyleService } from '../../../services';
+import { NumberService, JSONService, ObjectService, StringService } from '../../../services/utils';
 import DataStorage from '../../../services/storage';
 import { CustomWindowEvents, CustomEvents } from '../../../settings';
 import { styleErrors } from '../../../components/common/styles/errors';
@@ -10,7 +10,8 @@ import { BoolEnums } from '../../../enums';
 import { ContentSwSides, LabelIcons } from '../contentSw/enums';
 import { TextEditorHelper } from './textEditorHelper';
 import { PackIds } from '../../../theme/enums';
-import { FileSaveEnums, SaveEvts } from './enums';
+import { CustomMenuEvents } from './events';
+import { FileSaveEnums, SaveEvts, MenuButtons } from './enums';
 
 class TextEditor extends HTMLElement {
     constructor() {
@@ -38,6 +39,9 @@ class TextEditor extends HTMLElement {
             cols: 0
         };
 
+        this.content = '';
+        this.previewContent = '';
+        this.previewToggled = false;
         this.storage = new DataStorage();
         this.setfiles();
     }
@@ -48,8 +52,6 @@ class TextEditor extends HTMLElement {
         this.$loading = IdService.id('loading', this.shadow);
         this.$error = IdService.id('error', this.shadow);
         this.$cntSwitcher = IdService.id(this.contentSwId, this.shadow);
-        this.$textName = IdService.id(this.nameId, this.shadow);
-        this.$textArea = IdService.id(this.textAreaId, this.shadow);
 
         this.activateFile(0);
         this.saveFile(); // load real labels
@@ -64,8 +66,11 @@ class TextEditor extends HTMLElement {
         });
 
         CustomEventService.event(`${CustomEvents.interaction.textAreaChange}-${this.textAreaId}`, (e) => {
+            this.toggleContent(e.detail.value);
             this.saveFile(SaveEvts.content, e.detail.value);
         });
+
+        this.setMenuHandlera();
     }
 
     disconnectedCallback() {
@@ -76,24 +81,77 @@ class TextEditor extends HTMLElement {
         ]);
     }
 
+    setMenuHandlera() {
+        CustomEventService.event(`${CustomMenuEvents.menuClick}-${MenuButtons.paragraph.id}`, () => {
+            this.setTextTags('p');
+         });
+        CustomEventService.event(`${CustomMenuEvents.menuClick}-${MenuButtons.bold.id}`, () => {
+           this.setTextTags('b');
+        });
+        CustomEventService.event(`${CustomMenuEvents.menuClick}-${MenuButtons.italic.id}`, () => {
+            this.setTextTags('i');
+        });
+        CustomEventService.event(`${CustomMenuEvents.menuClick}-${MenuButtons.preview.id}`, () => {
+            HTMLService.html(this.$preview, this.previewContent);
+            this.toggleMode();
+        });
+    }
+
+    setTextTags(tag) {
+        let selection = StringService.getSelectedText();
+        selection = selection ? selection.toString() : '';
+        let html = this.content;
+        let selectionWithTag = `<${tag}>${selection}</${tag}>`;
+        html = html.replace(selection, selectionWithTag);
+        this.$textArea.setAttribute('value', html);
+        this.saveFile(SaveEvts.content, html);
+        this.togglePreviewContent(html);
+    }
+
     saveFile(evt, value) {  
         const saved = this.storage.getObject(FileSaveEnums.object);
         TextEditorHelper.saveFile(evt, value, saved, this.textObject, this.sets, this.$error, (savedArray) => {
             this.storage.saveObject(FileSaveEnums.object, savedArray);
             this.updateLabels(savedArray);
+            if (evt === SaveEvts.content) {
+                this.togglePreviewContent(value);
+            }
         });
     }
 
     getFileObject(file) {
         const saved = this.storage.getObject(FileSaveEnums.object);
         this.textObject = TextEditorHelper.getFileObject(saved, file);
+        if (this.textObject && this.textObject.content) {
+            this.toggleContent(this.textObject.content);
+            this.togglePreviewContent(this.textObject.content);
+        }
 
         this.toggleContentLoaded(false);
         TextEditorHelper.setLoading(NumberService.randomInteger(1, 2), file.name, this.$loading, () => {
             let html = this.setEditor(this.textObject, this.sessionId); 
             HTMLService.html(this.$container, html);
+            this.$textName = IdService.id(this.nameId, this.shadow);
+            this.$textArea = IdService.id(this.textAreaId, this.shadow);
+            this.$edit = IdService.id('edit', this.shadow);
+            this.$preview = IdService.id('preview', this.shadow);
             this.toggleContentLoaded(true); 
         });
+    }
+
+    toggleContent(toggle) {
+        this.content = toggle;
+    }
+
+    togglePreviewContent(toggle) {
+        this.previewContent = toggle;
+    }
+
+    toggleMode() {
+        StyleService.setDisplayMultiple([this.$preview, this.$edit], false);
+        let el = this.previewToggled ? this.$preview : this.$edit;
+        StyleService.setDisplay(el, true);
+        this.previewToggled = !this.previewToggled;
     }
 
     updateLabels(savedArray) {
@@ -136,22 +194,36 @@ class TextEditor extends HTMLElement {
     }
 
     setEditor(file, sessionId) {
+        let menu = '';
+        if (this.sets.menu.enabled) {
+            menu = `<div>
+                <texteditor-menu
+                  id="textMenu"
+                  setsId="${this.setsId}">
+                </texteditor-menu>
+            </div>`;
+        }
         return `
-            <text-input
-                id="${this.nameId}" 
-                label="File name"
-                class-name="input-normal"
-                value="${file.name}"
-                type="text"           
-            >
-            </text-input>
-            <text-area-input id="${this.textAreaId}" 
-                name="editorView-${sessionId}"
-                value="${file.content}"
-                rows="${file.rows}" 
-                cols="${file.cols}"
-            >
-            </text-area-input>
+            ${menu}
+            <div id="edit" class="contents">
+                <text-input
+                    id="${this.nameId}" 
+                    label="File name"
+                    class-name="input-normal"
+                    value="${file.name}"
+                    type="text"           
+                >
+                </text-input>
+                <div class="textpad"> &nbsp; </div>
+                <text-area-input id="${this.textAreaId}" 
+                    name="editorView-${sessionId}"
+                    value="${file.content}"
+                    rows="${file.rows}" 
+                    cols="${file.cols}"
+                >
+                </text-area-input>
+            </div>    
+            <div id="preview"></div>
         `;
     }
 
@@ -202,6 +274,19 @@ class TextEditor extends HTMLElement {
                 min-width: 500px;
                 min-height: 500px;
               }  
+
+              .contents {
+                margin-top: 8px;
+              }
+              .textpad {
+                height: 6px;
+              }
+
+              #preview {
+                width: 90%;
+                border: 1px solid black;
+                background-color: white;
+              }
             </style>
             <div class="editor-wrapper">
                 <div id="error"></div>