+++ src/components/ui/textEditor/index.js
@@ -1,1 +1,3 @@
+import './menu/menuButton';
+import './menu/menu';
 import './editor';