+++ src/components/ui/textEditor/theme.js
@@ -12,6 +12,15 @@ export const TextFileViewerTheme = {
             background: 'white',
             text: 'black',
             border: 'blue',
-        }
+        },
+        menu: {
+            background: '#dcdcdc',
+            item: {
+                bck: 'white',
+                border: 'blue',
+                tipBg: 'black',
+                tipText: 'white'
+            }
+        },
     }
 };